// src/app/shared/services/interfaces/request.interface.ts
export interface Request {
    id?:string;
    serviceName: string;
    category: string;
    endDate: string;
    startDate: string;
    serviceId: string;
    requestText: string;
    status: string;
    userId: string;
    userName: string;
    workerId: string;
    workerName: string;
    startTime: string;
    endTime: string;
}